+++ api/public/stack.php
@@ -31,43 +31,18 @@ require_login();
 ?>
 <html>
   <head>
-    <meta charset="utf-8"/>
+    <? readfile(__DIR__ . '/stackhead.html') ?>
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/codemirror/5.31.0/codemirror.min.css" />
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/codemirror/5.31.0/addon/lint/lint.min.css" />
-    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.7.1/jquery.min.js"></script>
-    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/js/bootstrap.min.js"></script>
-    <link rel="stylesheet" href="cors.php?name=sortable.min.css" />
-    <link rel="stylesheet" href="../api.css" />
-    <script type="text/x-mathjax-config">
-      MathJax.Hub.Config({
-        tex2jax: {
-          inlineMath: [['$', '$'], ['\\[','\\]'], ['\\(','\\)']],
-          displayMath: [['$$', '$$']],
-          processEscapes: true,
-          skipTags: ["script","noscript","style","textarea","pre","code","button"]
-        },
-        showMathMenu: false
-      })
-    </script>
-    <script
-      type="text/javascript"
-      src="//cdnjs.cloudflare.com/ajax/libs/mathjax/2.7.9/MathJax.js?config=TeX-AMS_HTML"
-    >
-    </script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/js-yaml/3.10.0/js-yaml.min.js"></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/codemirror/5.31.0/codemirror.js"></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/codemirror/5.31.0/addon/lint/lint.js"></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/codemirror/5.31.0/addon/lint/yaml-lint.js"></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/codemirror/5.31.0/mode/yaml/yaml.js"></script>
-    <script src="stackjsvle.js"></script>
   </head>
   <body>
+    <? require_once(__DIR__ . '/stackshared.php'); ?>
     <script>
-      const timeOutHandler = new Object();
-      const inputPrefix = 'stackapi_input_';
-      const feedbackPrefix = 'stackapi_fb_';
-      const validationPrefix = 'stackapi_val_';
-
       // Create data for call to API.
       function collectData() {
         const res = {
@@ -78,302 +53,6 @@ require_login();
           readOnly: document.getElementById('readOnly').checked,
         };
         return res;
-      }
-
-      // Get the different input elements by tag and return object with values.
-      function collectAnswer() {
-        const inputs = document.getElementsByTagName('input');
-        const textareas = document.getElementsByTagName('textarea');
-        const selects = document.getElementsByTagName('select');
-        let res = {};
-        res = processNodes(res, inputs);
-        res = processNodes(res, textareas);
-        res = processNodes(res, selects);
-        return res;
-      }
-
-      // Return object of values of valid entries in an HTMLCollection.
-      function processNodes(res, nodes) {
-        for (let i = 0; i < nodes.length; i++) {
-          const element = nodes[i];
-          if (element.name.indexOf(inputPrefix) === 0 && element.name.indexOf('_val') === -1) {
-            if (element.type === 'checkbox' || element.type === 'radio') {
-              if (element.checked) {
-                res[element.name.slice(inputPrefix.length)] = element.value;
-              }
-            } else {
-              res[element.name.slice(inputPrefix.length)] = element.value;
-            }
-          }
-        }
-        return res;
-      }
-
-      // Display rendered question and solution.
-      function send() {
-        const http = new XMLHttpRequest();
-        const url = window.location.origin + '/render';
-        http.open("POST", url, true);
-        http.setRequestHeader('Content-Type', 'application/json');
-        http.onreadystatechange = function() {
-          if(http.readyState == 4) {
-            try {
-              const json = JSON.parse(http.responseText);
-              if (json.message) {
-                document.getElementById('errors').innerText = json.message;
-                return;
-              } else {
-                document.getElementById('errors').innerText = '';
-              }
-              renameIframeHolders();
-              let question = json.questionrender;
-              const inputs = json.questioninputs;
-              let correctAnswers = '';
-              // Show correct answers.
-              for (const [name, input] of Object.entries(inputs)) {
-                question = question.replace(`[[input:${name}]]`, input.render);
-                question = question.replace(`[[validation:${name}]]`, `<span name='${validationPrefix + name}'></span>`);
-                if (input.samplesolutionrender && name !== 'remember') {
-                  // Display render of answer and matching user input to produce the answer.
-                  correctAnswers += `<p>
-                    <?php echo stack_string('api_correct_answer')?> \\[{${input.samplesolutionrender}}\\],
-                    <?php echo stack_string('api_which_typed')?>: `;
-                  for (const [name, solution] of Object.entries(input.samplesolution)) {
-                    if (name.indexOf('_val') === -1) {
-                      correctAnswers += `<span class='correct-answer'>${solution}</span>`;
-                    }
-                  }
-                  correctAnswers += '.</p>';
-                } else if (name !== 'remember') {
-                  // For dropdowns, radio buttons, etc, only the correct option is displayed.
-                  for (const solution of Object.values(input.samplesolution)) {
-                    if (input.configuration.options) {
-                      correctAnswers += `<p class='correct-answer'>${input.configuration.options[solution]}</p>`;
-                    }
-                  }
-                }
-              }
-              // Convert Moodle plot filenames to API filenames.
-              for (const [name, file] of Object.entries(json.questionassets)) {
-                question = question.replace(name, `plots/${file}`);
-                json.questionsamplesolutiontext = json.questionsamplesolutiontext.replace(name, `plots/${file}`);
-                correctAnswers = correctAnswers.replace(name, `plots/${file}`);
-              }
-              question = replaceFeedbackTags(question);
-
-              document.getElementById('output').innerHTML = question;
-              // Only display results sections once question retrieved.
-              document.getElementById('stackapi_qtext').style.display = 'block';
-              document.getElementById('stackapi_correct').style.display = 'block';
-
-              // Setup a validation call on inputs. Timeout length is reset if the input is updated
-              // before the validation call is made.
-              for (const inputName of Object.keys(inputs)) {
-                const inputElements = document.querySelectorAll(`[name^=${inputPrefix + inputName}]`);
-                for (const inputElement of Object.values(inputElements)) {
-                  inputElement.oninput = (event) => {
-                    const currentTimeout = timeOutHandler[event.target.id];
-                    if (currentTimeout) {
-                      window.clearTimeout(currentTimeout);
-                    }
-                    timeOutHandler[event.target.id] = window.setTimeout(validate.bind(null, event.target), 1000);
-                  };
-                }
-              }
-              let sampleText = json.questionsamplesolutiontext;
-              if (sampleText) {
-                sampleText = replaceFeedbackTags(sampleText);
-                document.getElementById('generalfeedback').innerHTML = sampleText;
-                document.getElementById('stackapi_generalfeedback').style.display = 'block';
-              } else {
-                // If the question is updated, there may no longer be general feedback.
-                document.getElementById('stackapi_generalfeedback').style.display = 'none';
-              }
-              document.getElementById('stackapi_score').style.display = 'none';
-              document.getElementById('stackapi_validity').innerText = '';
-              const innerFeedback = document.getElementById('specificfeedback');
-              innerFeedback.innerHTML = '';
-              innerFeedback.classList.remove('feedback');
-              document.getElementById('formatcorrectresponse').innerHTML = correctAnswers;
-              createIframes(json.iframes);
-              MathJax.Hub.Queue(["Typeset", MathJax.Hub]);
-            }
-            catch(e) {
-              document.getElementById('errors').innerText = http.responseText;
-              return;
-            }
-          }
-        };
-        http.send(JSON.stringify(collectData()));
-      }
-
-      // Validate an input. Called a set amount of time after an input is last updated.
-      function validate(element) {
-        const http = new XMLHttpRequest();
-        const url = window.location.origin + '/validate';
-        http.open("POST", url, true);
-        // Remove API prefix and subanswer id.
-        const answerName = element.name.slice(15).split('_', 1)[0];
-        http.setRequestHeader('Content-Type', 'application/json');
-        http.onreadystatechange = function() {
-          if(http.readyState == 4) {
-            try {
-              const json = JSON.parse(http.responseText);
-              if (json.message) {
-                document.getElementById('errors').innerText = json.message;
-                return;
-              } else {
-                document.getElementById('errors').innerText = '';
-              }
-              renameIframeHolders();
-              const validationHTML = json.validation;
-              const element = document.getElementsByName(`${validationPrefix + answerName}`)[0];
-              element.innerHTML = validationHTML;
-              if (validationHTML) {
-                element.classList.add('validation');
-              } else {
-                element.classList.remove('validation');
-              }
-              createIframes(json.iframes);
-              MathJax.Hub.Queue(["Typeset", MathJax.Hub]);
-            }
-            catch(e) {
-              document.getElementById('errors').innerText = http.responseText;
-              return;
-            }
-          }
-        };
-
-        const data = collectData();
-        data.inputName = answerName;
-        http.send(JSON.stringify(data));
-      }
-
-      // Submit answers.
-      function answer() {
-        const http = new XMLHttpRequest();
-        const url = window.location.origin + '/grade';
-        http.open("POST", url, true);
-
-        if (!document.getElementById('output').innerText) {
-          return;
-        }
-
-        http.setRequestHeader('Content-Type', 'application/json');
-        http.onreadystatechange = function() {
-          if(http.readyState == 4) {
-            try {
-              const json = JSON.parse(http.responseText);
-              if (json.message) {
-                document.getElementById('errors').innerText = json.message;
-                return;
-              } else {
-                document.getElementById('errors').innerText = '';
-              }
-              if (!json.isgradable) {
-                document.getElementById('stackapi_validity').innerText
-                  = ' <?php echo stack_string('api_valid_all_parts')?>';
-                return;
-              }
-              renameIframeHolders();
-              document.getElementById('score').innerText
-                = (json.score * json.scoreweights.total).toFixed(2) +
-                ' <?php echo stack_string('api_out_of')?> ' + json.scoreweights.total;
-              document.getElementById('stackapi_score').style.display = 'block';
-              document.getElementById('response_summary').innerText = json.responsesummary;
-              document.getElementById('stackapi_summary').style.display = 'block';
-              const feedback = json.prts;
-              const specificFeedbackElement = document.getElementById('specificfeedback');
-              // Replace tags and plots in specific feedback and then display.
-              if (json.specificfeedback) {
-                for (const [name, file] of Object.entries(json.gradingassets)) {
-                  json.specificfeedback = json.specificfeedback.replace(name, `plots/${file}`);
-                }
-                json.specificfeedback = replaceFeedbackTags(json.specificfeedback);
-                specificFeedbackElement.innerHTML = json.specificfeedback;
-                specificFeedbackElement.classList.add('feedback');
-              } else {
-                specificFeedbackElement.classList.remove('feedback');
-              }
-              // Replace plots in tagged feedback and then display.
-              for (let [name, fb] of Object.entries(feedback)) {
-                for (const [name, file] of Object.entries(json.gradingassets)) {
-                  fb = fb.replace(name, `plots/${file}`);
-                }
-                const elements = document.getElementsByName(`${feedbackPrefix + name}`);
-                if (elements.length > 0) {
-                  const element = elements[0];
-                  if (json.scores[name] !== undefined) {
-                    fb = fb + `<div><?php echo stack_string('api_marks_sub')?>:
-                      ${(json.scores[name] * json.scoreweights[name] * json.scoreweights.total).toFixed(2)}
-                        / ${(json.scoreweights[name] * json.scoreweights.total).toFixed(2)}.</div>`;
-                  }
-                  element.innerHTML = fb;
-                  if (fb) {
-                    element.classList.add('feedback');
-                  } else {
-                    element.classList.remove('feedback');
-                  }
-                }
-              }
-              createIframes(json.iframes);
-              MathJax.Hub.Queue(["Typeset", MathJax.Hub]);
-            }
-            catch(e) {
-              document.getElementById('errors').innerText = http.responseText;
-              return;
-            }
-          }
-        };
-        // Clear previous answers and score.
-        const specificFeedbackElement = document.getElementById('specificfeedback');
-        specificFeedbackElement.innerHTML = "";
-        specificFeedbackElement.classList.remove('feedback');
-        document.getElementById('response_summary').innerText = "";
-        document.getElementById('stackapi_summary').style.display = 'none';
-        const inputElements = document.querySelectorAll(`[name^=${feedbackPrefix}]`);
-        for (const inputElement of Object.values(inputElements)) {
-          inputElement.innerHTML = "";
-          inputElement.classList.remove('feedback');
-        }
-        document.getElementById('stackapi_score').style.display = 'none';
-        document.getElementById('stackapi_validity').innerText = '';
-        http.send(JSON.stringify(collectData()));
-      }
-
-      function download(filename, fileid) {
-        const http = new XMLHttpRequest();
-        const url = window.location.origin + '/download';
-        http.open("POST", url, true);
-        http.setRequestHeader('Content-Type', 'application/json');
-        // Something funky going on with closures and callbacks. This seems
-        // to be the easiest way to pass through the file details.
-        http.filename = filename;
-        http.fileid = fileid;
-        http.onreadystatechange = function() {
-          if(http.readyState == 4) {
-            try {
-              // Only download the file once. Replace call to download controller with link
-              // to downloaded file.
-              const blob = new Blob([http.responseText], {type: 'application/octet-binary', endings: 'native'});
-              const selector = CSS.escape(`javascript\:download\(\'${http.filename}\'\, ${http.fileid}\)`)
-              const linkElements = document.querySelectorAll(`a[href^=${selector}]`);
-
-              const link = linkElements[0];
-              link.setAttribute('href', URL.createObjectURL(blob));
-              link.setAttribute('download', filename);
-              link.click();
-            }
-            catch(e) {
-              document.getElementById('errors').innerText = http.responseText;
-              return;
-            }
-          }
-        };
-        const data = collectData();
-        data.filename = filename;
-        data.fileid = fileid;
-        http.send(JSON.stringify(data));
       }
 
       // Save contents of question editor locally.
@@ -389,33 +68,6 @@ require_login();
           return localStorage.getItem(key) || '';
         }
         return '';
-      }
-
-      function renameIframeHolders() {
-        // Each call to STACK restarts numbering of iframe holders so we need to rename
-        // any old ones to make sure new iframes end up in the correct place.
-        for (const iframe of document.querySelectorAll(`[id^=stack-iframe-holder]:not([id$=old]`)) {
-          iframe.id = iframe.id + '_old';
-        }
-      }
-
-      function createIframes (iframes) {
-        for (const iframe of iframes) {
-          create_iframe(...iframe);
-        }
-      }
-
-      // Replace feedback tags in some text with an approproately named HTML div.
-      function replaceFeedbackTags(text) {
-        let result = text;
-        const feedbackTags = text.match(/\[\[feedback:.*\]\]/g);
-        if (feedbackTags) {
-          for (const tag of feedbackTags) {
-            // Part name is between '[[feedback:' and ']]'.
-            result = result.replace(tag, `<div name='${feedbackPrefix + tag.slice(11, -2)}'></div>`);
-          }
-        }
-        return result;
       }
 
       function getQuestionFile(questionURL) {
@@ -464,7 +116,6 @@ require_login();
       function setQuestion(question) {
         yamlEditor.getDoc().setValue('<quiz>\n' + question + '\n</quiz>');
       }
-
     </script>
 
     <div class="container-fluid que stack">
