+++ api/public/stackhead.html
@@ -1,0 +1,22 @@
+<meta charset="utf-8"/>
+<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.7.1/jquery.min.js"></script>
+<link rel="stylesheet" href="../api.css" />
+<script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/js/bootstrap.min.js"></script>
+<link rel="stylesheet" href="cors.php?name=sortable.min.css" />
+<script type="text/x-mathjax-config">
+  MathJax.Hub.Config({
+    tex2jax: {
+      inlineMath: [['$', '$'], ['\\[','\\]'], ['\\(','\\)']],
+      displayMath: [['$$', '$$']],
+      processEscapes: true,
+      skipTags: ["script","noscript","style","textarea","pre","code","button"]
+    },
+    showMathMenu: false
+  })
+</script>
+<script
+  type="text/javascript"
+  src="//cdnjs.cloudflare.com/ajax/libs/mathjax/2.7.9/MathJax.js?config=TeX-AMS_HTML"
+>
+</script>
+<script src="stackjsvle.js"></script>
